decouple renderer from scripting
